Fix for state updates not being able to be edited
